wire: move the TABLE_DUMP_V2 next-hop decoder into the codec crate
The shared RIB-entry MP_REACH_NLRI next-hop decoder introduced on this
branch lived in rustbgpd-mrt, which depends on rustbgpd-rib and
rustbgpd-telemetry; linking it from the CLI pulled the RIB into every
rbgp build and grew the debug binary from 230 MB to 337 MB.

Move the decoder into rustbgpd-wire as
`mrt::decode_table_dump_v2_mp_reach_next_hop`, which both readers already
depend on and which owns MP_REACH_NLRI parsing. It returns
`DecodeError::MalformedField` for every rejected shape; the mrt reader
wraps that into `ReadError::Malformed` at the attribute value's absolute
offset and the CLI maps it into its string error. Drop the CLI's
rustbgpd-mrt dependency, restore its private TABLE_DUMP_V2 subtype
constants (registry values that cannot drift), and return the mrt crate's
constants to crate visibility.

Prepare rustbgpd-wire 0.19.1 for the additive API: bump the manifest and
workspace pin, record the release in the crate changelog and README
compatibility note, and add the RFC 6396 row and `mrt` module entry.
