+++ Cargo.toml
@@ -35,7 +35,7 @@ categories = ["network-programming"]
 
 [workspace.dependencies]
 # Internal crates
-rustbgpd-wire = { path = "crates/wire", version = "0.19.0" }
+rustbgpd-wire = { path = "crates/wire", version = "0.19.1" }
 rustbgpd-bfd = { path = "crates/bfd", version = "0.68.0" }
 rustbgpd-fsm = { path = "crates/fsm", version = "0.6.0" }
 rustbgpd-transport = { path = "crates/transport", version = "0.68.0" }
