+++ crates/mrt/src/reader.rs
@@ -9,7 +9,8 @@
 //! L2VPN/EVPN) are skipped and counted, not treated as errors. A RIB
 //! entry's `MP_REACH_NLRI` is accepted in both the RFC 6396 §4.3.4
 //! reduced form the writer emits and the full RFC 4760 form other
-//! collectors write; see [`decode_rib_entry_mp_reach_next_hop`].
+//! collectors write; see [`decode_table_dump_v2_mp_reach_next_hop`] in
+//! `rustbgpd_wire`.
 //!
 //! Input is treated as hostile — this will read an untrusted file at
 //! daemon boot. Every length field is bounds-checked, malformed input
@@ -30,6 +31,7 @@ use bytes::Bytes;
 use rustbgpd_rib::update::MrtPeerEntry;
 use rustbgpd_wire::constants::{attr_flags, attr_type};
 use rustbgpd_wire::error::DecodeError;
+use rustbgpd_wire::mrt::decode_table_dump_v2_mp_reach_next_hop;
 use rustbgpd_wire::notification::update_subcode;
 use rustbgpd_wire::{Afi, ErrorDisposition, Ipv4Prefix, Ipv6Prefix, PathAttribute, Prefix, Safi};
 use std::borrow::Cow;
@@ -265,101 +267,6 @@ fn read_record(data: &[u8], start: usize) -> Result<(u16, u16, &[u8]), ReadError
     Ok((mrt_type, subtype, &data[payload_start..payload_start + len]))
 }
 
-/// IANA AFI numbers accepted in a full-form `MP_REACH_NLRI` header.
-const AFI_IPV4: u16 = Afi::Ipv4 as u16;
-const AFI_IPV6: u16 = Afi::Ipv6 as u16;
-
-/// Decode the `MP_REACH_NLRI` value of a `TABLE_DUMP_V2` RIB entry into
-/// its next hop, plus the link-local address when the next hop is the
-/// 32-octet global + link-local pair.
-///
-/// RFC 6396 §4.3.4 reduces the attribute to next-hop length (1 octet) and
-/// next hop, but several collectors write the full RFC 4760 form (AFI,
-/// SAFI, next-hop length, next hop, reserved octet). Both are accepted,
-/// told apart by the first octet: a next-hop length is never 0, so a
-/// leading 0 can only be the high octet of AFI 1 or 2 — the discriminator
-/// bgpdump, bgpkit-parser, and mrtparse also use. Everything else is
-/// rejected: a next-hop length other than 4, 16, or 32; a truncated next
-/// hop; a full-form AFI that disagrees with the next-hop length (AFI 1
-/// requires 4, AFI 2 requires 16 or 32); and any octets past the next hop
-/// beyond the single reserved octet the full form may carry (its value is
-/// ignored, RFC 4760 §3).
-///
-/// Shared by the daemon's warm-checkpoint reader and `rbgp diff snapshot
-/// from-mrt`, so one dump converts and reads back identically.
-///
-/// # Errors
-///
-/// [`ReadError::Malformed`] for every rejected shape; its `offset` is
-/// relative to the start of `value`.
-pub fn decode_rib_entry_mp_reach_next_hop(
-    value: &[u8],
-) -> Result<(IpAddr, Option<Ipv6Addr>), ReadError> {
-    let Some(&first) = value.first() else {
-        return Err(malformed(0, "empty MP_REACH_NLRI value"));
-    };
-    let (afi, nh_len_at) = if first == 0 {
-        if value.len() < 4 {
-            return Err(malformed(
-                value.len(),
-                "truncated full-form MP_REACH_NLRI header (AFI, SAFI, NH-Len)",
-            ));
-        }
-        (Some(u16::from_be_bytes([value[0], value[1]])), 3)
-    } else {
-        (None, 0)
-    };
-    let nh_len = usize::from(value[nh_len_at]);
-    let rest = &value[nh_len_at + 1..];
-    let v6 = |o: &[u8; 16]| IpAddr::V6(Ipv6Addr::from(*o));
-    let decoded = match (afi, nh_len) {
-        (None | Some(AFI_IPV4), 4) => rest
-            .split_first_chunk::<4>()
-            .map(|(o, tail)| ((IpAddr::V4(Ipv4Addr::from(*o)), None), tail)),
-        (None | Some(AFI_IPV6), 16) => rest
-            .split_first_chunk::<16>()
-            .map(|(o, tail)| ((v6(o), None), tail)),
-        (None | Some(AFI_IPV6), 32) => rest.split_first_chunk::<16>().and_then(|(g, tail)| {
-            tail.split_first_chunk::<16>()
-                .map(|(ll, tail)| ((v6(g), Some(Ipv6Addr::from(*ll))), tail))
-        }),
-        (None, _) => {
-            return Err(malformed(
-                nh_len_at,
-                format!("MP_REACH_NLRI NH-Len {nh_len} is not 4, 16, or 32"),
-            ));
-        }
-        (Some(afi), _) => {
-            return Err(malformed(
-                nh_len_at,
-                format!("MP_REACH_NLRI AFI {afi} does not carry a {nh_len}-octet next hop"),
-            ));
-        }
-    };
-    let Some((next_hop, tail)) = decoded else {
-        return Err(malformed(
-            value.len(),
-            format!(
-                "truncated MP_REACH_NLRI next hop: NH-Len {nh_len}, {} octets present",
-                rest.len()
-            ),
-        ));
-    };
-    // The full form ends with one reserved octet (RFC 4760 §3); nothing
-    // else may follow the next hop in either form.
-    let reserved = usize::from(afi.is_some());
-    if tail.len() > reserved {
-        return Err(malformed(
-            value.len() - tail.len() + reserved,
-            format!(
-                "{} trailing octets after MP_REACH_NLRI next hop",
-                tail.len() - reserved
-            ),
-        ));
-    }
-    Ok(next_hop)
-}
-
 /// Decoded attribute block of one RIB entry.
 struct EntryAttributes {
     attributes: Vec<PathAttribute>,
@@ -411,14 +318,10 @@ fn decode_entry_attributes(attrs: &[u8], base: usize) -> Result<EntryAttributes,
                     },
                 ));
             }
-            mp_next_hop = Some(decode_rib_entry_mp_reach_next_hop(value).map_err(
-                |err| match err {
-                    ReadError::Malformed { offset, context } => {
-                        malformed(value_offset + offset, context)
-                    }
-                    other => other,
-                },
-            )?);
+            mp_next_hop = Some(
+                decode_table_dump_v2_mp_reach_next_hop(value)
+                    .map_err(|err| malformed(value_offset, err.to_string()))?,
+            );
         } else {
             other.extend_from_slice(&attrs[span_start..cur.rel_pos()]);
         }
@@ -1642,98 +1545,8 @@ mod tests {
         );
     }
 
-    fn cat(parts: &[&[u8]]) -> Vec<u8> {
-        parts.concat()
-    }
-
-    #[test]
-    fn rib_entry_mp_reach_accepts_reduced_and_full_forms() {
-        let global = Ipv6Addr::new(0x2001, 0xdb8, 0, 0, 0, 0, 0, 1);
-        let link_local = Ipv6Addr::new(0xfe80, 0, 0, 0, 0, 0, 0, 1);
-        let v4 = [192, 0, 2, 1];
-        let v6 = global.octets();
-        let pair = cat(&[&v6, &link_local.octets()]);
-        let cases = [
-            // RFC 6396 §4.3.4 reduced form: NH-Len + next hop.
-            (cat(&[&[4], &v4]), (IpAddr::V4(v4.into()), None)),
-            (cat(&[&[16], &v6]), (IpAddr::V6(global), None)),
-            (cat(&[&[32], &pair]), (IpAddr::V6(global), Some(link_local))),
-            // Full RFC 4760 form: AFI, SAFI, NH-Len, next hop, optional
-            // reserved octet.
-            (
-                cat(&[&[0, 1, 1, 4], &v4, &[0]]),
-                (IpAddr::V4(v4.into()), None),
-            ),
-            (cat(&[&[0, 1, 1, 4], &v4]), (IpAddr::V4(v4.into()), None)),
-            (
-                cat(&[&[0, 2, 1, 16], &v6, &[0]]),
-                (IpAddr::V6(global), None),
-            ),
-            (cat(&[&[0, 2, 1, 16], &v6]), (IpAddr::V6(global), None)),
-            (
-                cat(&[&[0, 2, 1, 32], &pair, &[0]]),
-                (IpAddr::V6(global), Some(link_local)),
-            ),
-        ];
-        for (value, expected) in cases {
-            assert_eq!(
-                decode_rib_entry_mp_reach_next_hop(&value).unwrap(),
-                expected,
-                "value {value:?}"
-            );
-        }
-    }
-
-    #[test]
-    fn rib_entry_mp_reach_rejects_malformed_shapes() {
-        let v4 = [192, 0, 2, 1];
-        let v6 = Ipv6Addr::new(0x2001, 0xdb8, 0, 0, 0, 0, 0, 1).octets();
-        // (value, expected error text, expected offset within value)
-        let rejected = [
-            (vec![], "empty", 0),
-            // A leading 0 is an AFI high octet, so a lone 0 (or a
-            // reduced-form NH-Len 0) is a truncated full-form header.
-            (vec![0], "truncated full-form", 1),
-            (vec![0, 0, 0, 0], "AFI 0 does not carry a 0-octet", 3),
-            (cat(&[&[5], &[1, 2, 3, 4, 5]]), "NH-Len 5 is not", 0),
-            (vec![16, 0, 0], "truncated MP_REACH_NLRI next hop", 3),
-            (
-                cat(&[&[0, 2, 1, 16], &[0, 0]]),
-                "truncated MP_REACH_NLRI next hop",
-                6,
-            ),
-            // The reduced form has no reserved octet.
-            (cat(&[&[4], &v4, &[0]]), "1 trailing octets", 5),
-            (cat(&[&[0, 1, 1, 4], &v4, &[0, 0]]), "1 trailing octets", 9),
-            (
-                cat(&[&[0, 2, 1, 4], &v4, &[0]]),
-                "AFI 2 does not carry a 4-octet",
-                3,
-            ),
-            (
-                cat(&[&[0, 1, 1, 16], &v6, &[0]]),
-                "AFI 1 does not carry a 16-octet",
-                3,
-            ),
-            (
-                cat(&[&[0, 2, 1, 5], &[1, 2, 3, 4, 5]]),
-                "AFI 2 does not carry a 5-octet",
-                3,
-            ),
-        ];
-        for (value, needle, expected_offset) in rejected {
-            match decode_rib_entry_mp_reach_next_hop(&value) {
-                Err(ReadError::Malformed { offset, context }) => {
-                    assert!(context.contains(needle), "value {value:?}: {context}");
-                    assert_eq!(offset, expected_offset, "value {value:?}: {context}");
-                }
-                other => panic!("value {value:?}: expected Malformed, got {other:?}"),
-            }
-        }
-    }
-
     /// The entry-level error carries the absolute file offset of the
-    /// offending octet, not one relative to the attribute value.
+    /// attribute value the shared wire decoder rejected.
     #[test]
     fn rib_entry_mp_reach_trailing_octets_fail_the_entry_at_absolute_offset() {
         let peer = make_peer(IpAddr::V4(Ipv4Addr::new(10, 0, 0, 1)), 65001);
@@ -1760,7 +1573,7 @@ mod tests {
             panic!("expected Malformed, got {err:?}");
         };
         assert!(context.contains("1 trailing octets"), "{context}");
-        assert_eq!(offset, value_offset + 5);
+        assert_eq!(offset, value_offset);
     }
 
     #[test]
